chore(Comix.to): "utils.ts" structure
